+++ src/commands/mirror_node.ts
@@ -15,7 +15,7 @@
  *
  */
 import {ListrEnquirerPromptAdapter} from '@listr2/prompt-adapter-enquirer';
-import {Listr} from 'listr2';
+import {Listr, type ListrTask} from 'listr2';
 import {SoloError, IllegalArgumentError, MissingArgumentError} from '../core/errors.js';
 import * as constants from '../core/constants.js';
 import {type AccountManager} from '../core/account_manager.js';
@@ -25,8 +25,10 @@ import {Flags as flags} from './flags.js';
 import {getEnvValue} from '../core/helpers.js';
 import {RemoteConfigTasks} from '../core/config/remote/remote_config_tasks.js';
 import {type CommandBuilder, type PodName} from '../types/aliases.js';
-import type {Opts} from '../types/index.js';
+import {type Opts} from '../types/command_types.js';
 import {ListrLease} from '../core/lease/listr_lease.js';
+import {ComponentType} from '../core/config/remote/enumerations.js';
+import {MirrorNodeComponent} from '../core/config/remote/components/mirror_node_component.js';
 
 export class MirrorNodeCommand extends BaseCommand {
   private readonly accountManager: AccountManager;
@@ -421,7 +423,7 @@ export class MirrorNodeCommand extends BaseCommand {
             );
           },
         },
-        RemoteConfigTasks.addMirrorNodeAndMirrorNodeToExplorer.bind(this)(),
+        MirrorNodeCommand.addMirrorNodeAndMirrorNodeToExplorer.bind(this)(),
       ],
       {
         concurrent: false,
@@ -518,7 +520,7 @@ export class MirrorNodeCommand extends BaseCommand {
           },
           skip: ctx => !ctx.config.isChartInstalled,
         },
-        RemoteConfigTasks.removeMirrorNodeAndMirrorNodeToExplorer.bind(this)(),
+        MirrorNodeCommand.removeMirrorNodeAndMirrorNodeExplorer.bind(this)(),
       ],
       {
         concurrent: false,
@@ -588,6 +590,44 @@ export class MirrorNodeCommand extends BaseCommand {
             },
           })
           .demandCommand(1, 'Select a mirror-node command');
+      },
+    };
+  }
+
+  /** Removes the mirror node and mirror node explorer components from remote config. */
+  public static removeMirrorNodeAndMirrorNodeExplorer(this: MirrorNodeCommand): ListrTask<any, any, any> {
+    return {
+      title: 'Remove mirror node and mirror node explorer from remote config',
+      skip: (): boolean => !this.remoteConfigManager.isLoaded(),
+      task: async (): Promise<void> => {
+        await this.remoteConfigManager.modify(async remoteConfig => {
+          remoteConfig.components.remove('mirrorNode', ComponentType.MirrorNode);
+
+          remoteConfig.components.remove('mirrorNodeExplorer', ComponentType.MirrorNode);
+        });
+      },
+    };
+  }
+
+  /** Adds the mirror node and mirror node explorer components to remote config. */
+  public static addMirrorNodeAndMirrorNodeToExplorer(this: MirrorNodeCommand): ListrTask<any, any, any> {
+    return {
+      title: 'Add mirror node and mirror node explorer to remote config',
+      skip: (): boolean => !this.remoteConfigManager.isLoaded(),
+      task: async (ctx): Promise<void> => {
+        await this.remoteConfigManager.modify(async remoteConfig => {
+          const {
+            config: {namespace},
+          } = ctx;
+          const cluster = this.remoteConfigManager.currentCluster;
+
+          remoteConfig.components.add('mirrorNode', new MirrorNodeComponent('mirrorNode', cluster, namespace));
+
+          remoteConfig.components.add(
+            'mirrorNodeExplorer',
+            new MirrorNodeComponent('mirrorNodeExplorer', cluster, namespace),
+          );
+        });
       },
     };
   }
