+++ src/commands/relay.ts
@@ -25,7 +25,7 @@ import {Flags as flags} from './flags.js';
 import {getNodeAccountMap} from '../core/helpers.js';
 import {RemoteConfigTasks} from '../core/config/remote/remote_config_tasks.js';
 import {type CommandBuilder, type NodeAliases} from '../types/aliases.js';
-import {type Opts} from '../types/index.js';
+import {type Opts} from '../types/command_types.js';
 import {ListrLease} from '../core/lease/listr_lease.js';
 
 export class RelayCommand extends BaseCommand {
