+++ src/core/templates.ts
@@ -22,7 +22,7 @@ import * as constants from './constants.js';
 import {type AccountId} from '@hashgraph/sdk';
 import type {NodeAlias, PodName} from '../types/aliases.js';
 import {GrpcProxyTlsEnums} from './enumerations.js';
-import type {ContextClusterStructure} from '../types/index.js';
+import {type ContextClusterStructure} from '../types/config_types.js';
 import type {Cluster, Context} from './config/remote/types.js';
 import {Flags as flags} from '../commands/flags.js';
