+++ src/core/yargs_command.ts
@@ -17,7 +17,7 @@
 import {Flags as commandFlags} from '../commands/flags.js';
 import {IllegalArgumentError} from './errors.js';
 import {type BaseCommand} from '../commands/base.js';
-import {type CommandFlag} from '../types/index.js';
+import {type CommandFlag} from '../types/flag_types.js';
 
 export class YargsCommand {
   constructor(
