+++ src/types/aliases.ts
@@ -31,4 +31,3 @@ export type TaskFunction = (
   task: ListrTaskWrapper<any, any, any>,
 ) => Promise<Listr<any, any, any>> | Listr<any, any, any> | Promise<void> | void;
 export type ConfigBuilder = (argv, ctx, task) => Promise<any>;
-export type PromptFunction = (task: ListrTaskWrapper<any, any, any>, input: any) => Promise<any>;
