+++ src/types/index.ts
@@ -18,24 +18,8 @@ import type * as x509 from '@peculiar/x509';
 import type net from 'net';
 import type * as WebSocket from 'ws';
 import type crypto from 'crypto';
-import type {SoloLogger} from '../core/logging.js';
-import {type PlatformInstaller} from '../core/platform_installer.js';
-import {type ChartManager} from '../core/chart_manager.js';
-import {type ConfigManager} from '../core/config_manager.js';
-import {type DependencyManager} from '../core/dependency_managers/index.js';
-import {type KeyManager} from '../core/key_manager.js';
-import {type AccountManager} from '../core/account_manager.js';
-import {type ProfileManager} from '../core/profile_manager.js';
-import {type LeaseManager} from '../core/lease/lease_manager.js';
-import {type CertificateManager} from '../core/certificate_manager.js';
-import {type LocalConfig} from '../core/config/local_config.js';
-import {type RemoteConfigManager} from '../core/config/remote/remote_config_manager.js';
-import {type Helm} from '../core/helm.js';
-import {type K8} from '../core/k8.js';
-import {type PackageDownloader} from '../core/package_downloader.js';
-import type {Cluster, Context} from '../core/config/remote/types.js';
-import {type BaseCommand} from '../commands/base.js';
-import {type PromptFunction} from './aliases.js';
+
+// NOTE: DO NOT add any Solo imports in this file to avoid circular dependencies
 
 export interface NodeKeyObject {
   privateKey: crypto.webcrypto.CryptoKey;
@@ -65,49 +49,10 @@ export interface AccountIdWithKeyPairObject {
   publicKey: string;
 }
 
-export interface CommandFlag {
-  constName: string;
-  name: string;
-  definition: Definition;
-  prompt: PromptFunction;
-}
-
-export interface Definition {
-  describe: string;
-  defaultValue?: boolean | string | number;
-  alias?: string;
-  type?: string;
-  disablePrompt?: boolean;
-}
-
-export interface Opts {
-  logger: SoloLogger;
-  helm: Helm;
-  k8: K8;
-  downloader: PackageDownloader;
-  platformInstaller: PlatformInstaller;
-  chartManager: ChartManager;
-  configManager: ConfigManager;
-  depManager: DependencyManager;
-  keyManager: KeyManager;
-  accountManager: AccountManager;
-  profileManager: ProfileManager;
-  leaseManager: LeaseManager;
-  certificateManager: CertificateManager;
-  localConfig: LocalConfig;
-  remoteConfigManager: RemoteConfigManager;
-}
-
-export interface CommandHandlers {
-  parent: BaseCommand;
-}
-
 /**
  * Generic type for representing optional types
  */
 export type Optional<T> = T | undefined;
-
-export type ContextClusterStructure = Record<Context, Cluster>;
 
 /**
  * Interface for capsuling validating for class's own properties
