+++ test/unit/core/remote_config/components.ts
@@ -17,15 +17,13 @@
 import {expect} from 'chai';
 import {describe, it} from 'mocha';
 
-import {
-  BaseComponent,
-  RelayComponent,
-  HaProxyComponent,
-  MirrorNodeComponent,
-  EnvoyProxyComponent,
-  ConsensusNodeComponent,
-  MirrorNodeExplorerComponent,
-} from '../../../../src/core/config/remote/components/index.js';
+import {RelayComponent} from '../../../../src/core/config/remote/components/relay_component.js';
+import {BaseComponent} from '../../../../src/core/config/remote/components/base_component.js';
+import {ConsensusNodeComponent} from '../../../../src/core/config/remote/components/consensus_node_component.js';
+import {HaProxyComponent} from '../../../../src/core/config/remote/components/ha_proxy_component.js';
+import {EnvoyProxyComponent} from '../../../../src/core/config/remote/components/envoy_proxy_component.js';
+import {MirrorNodeComponent} from '../../../../src/core/config/remote/components/mirror_node_component.js';
+import {MirrorNodeExplorerComponent} from '../../../../src/core/config/remote/components/mirror_node_explorer_component.js';
 import {SoloError} from '../../../../src/core/errors.js';
 import {ConsensusNodeStates} from '../../../../src/core/config/remote/enumerations.js';
 import type {NodeAliases} from '../../../../src/types/aliases.js';
